amended q.selector. Now - this.body = document.querySelector(#main-container) in notesView.js
