+++ notes-app/notesView.test.js
@@ -3,17 +3,17 @@
  */
 
 const fs = require("fs");
-const NotesView1 = require("./notesView1");
+const NotesView = require("./notesView");
 const NotesModel = require("./notesModel");
 
-describe("NotesView1", () => {
+describe("NotesView", () => {
   beforeEach(() => {
     document.body.innerHTML = fs.readFileSync("./index.html");
   });
 
   it("(1) creates a new div when notesView.displayNotes(); is called ", () => {
     const notesModel = new NotesModel();
-    const notesView = new NotesView1(notesModel);
+    const notesView = new NotesView(notesModel);
     notesModel.addNote("i'm a note!");
     notesView.displayNotes();
     expect(document.querySelectorAll("div").length).toBe(2);
@@ -21,7 +21,7 @@ describe("NotesView1", () => {
 
   it("(2) creates a new element with class name 'note'", () => {
     const notesModel = new NotesModel();
-    const notesView = new NotesView1(notesModel);
+    const notesView = new NotesView(notesModel);
     notesModel.addNote("i'm a note!");
     notesView.displayNotes();
     expect(document.getElementsByClassName("note").length).toBe(1);
@@ -29,7 +29,7 @@ describe("NotesView1", () => {
 
   it("(3) first element with class name 'note' is 'i'm a note!", () => {
     const notesModel = new NotesModel();
-    const notesView = new NotesView1(notesModel);
+    const notesView = new NotesView(notesModel);
     notesModel.addNote("i'm a note!");
     notesView.displayNotes();
     expect(document.getElementsByClassName("note")[0].innerText).toBe(
@@ -39,7 +39,7 @@ describe("NotesView1", () => {
 
   it("(4) tests 1,2,3 all pass in the same instance", () => {
     const notesModel = new NotesModel();
-    const notesView = new NotesView1(notesModel);
+    const notesView = new NotesView(notesModel);
     notesModel.addNote("i'm a note!");
     notesView.displayNotes();
     expect(document.getElementsByClassName("note").length).toBe(1);
@@ -51,7 +51,7 @@ describe("NotesView1", () => {
 
   it("(5) notesView.displayNotes() creates 2 divs", () => {
     const notesModel = new NotesModel();
-    const notesView = new NotesView1(notesModel);
+    const notesView = new NotesView(notesModel);
     notesModel.addNote("i'm a note!");
     notesModel.addNote("i'm a note 2!");
     notesModel.addNote("i'm a note 3!");
@@ -62,7 +62,7 @@ describe("NotesView1", () => {
 
   it("(6) notesView.displayNotes() creates 3 new divs with class name 'note'", () => {
     const notesModel = new NotesModel();
-    const notesView = new NotesView1(notesModel);
+    const notesView = new NotesView(notesModel);
     notesModel.addNote("i'm a note!");
     notesModel.addNote("i'm a note 2!");
     notesModel.addNote("i'm a note 3!");
@@ -75,4 +75,3 @@ describe("NotesView1", () => {
     );
   });
 });
-// document.getElementsByClassName("note").length &&
