+++ notes-app/notesView1.js
@@ -1,25 +1,18 @@
-class NotesView1 {
+class NotesView {
   constructor(model) {
     this.model = model;
-    this.body = document.querySelector("body");
-    this.newDiv = "";
+    this.mainContainerEl = document.querySelector("#main-container");
   }
 
   displayNotes() {
     const notes = this.model.getNotes();
-    // const newDiv = document.createElement("div");
     notes.forEach((note) => {
-      this.newDiv = document.createElement("div");
-      this.newDiv.classList.add("note");
-      this.newDiv.innerText = note;
-      this.body.append(this.newDiv);
+      const noteEl = document.createElement("div");
+      noteEl.textContent = note;
+      noteEl.className = "note";
+      this.mainContainerEl.append(noteEl);
     });
-
-    console.log(notes);
   }
 }
 
-module.exports = NotesView1;
-// document.getElementsByClassName("note");
-// return notes.forEach((newNote) => {
-//     this.mainContainerEl.append((newDiv.innerText = newNote));
+module.exports = NotesView;
